+++ tests/QueryPlannerTest.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Phlib\DbHelper\Tests;
 
 use Phlib\Db\Adapter;
@@ -20,16 +22,21 @@ class QueryPlannerTest extends TestCase
      */
     protected $adapter;
 
-    protected function setUp()
+    protected function setUp(): void
     {
         $this->adapter = $this->createMock(Adapter::class);
 
         parent::setUp();
     }
 
-    public function testGetPlanDoesExplain()
+    public function testGetPlanDoesExplain(): void
     {
         $pdoStatement = $this->createMock(\PDOStatement::class);
+        $pdoStatement->expects(static::once())
+            ->method('fetchAll')
+            ->with(\PDO::FETCH_ASSOC)
+            ->willReturn([]);
+
         $this->adapter->expects(static::once())
             ->method('query')
             ->with(static::stringContains('EXPLAIN', true))
@@ -38,7 +45,7 @@ class QueryPlannerTest extends TestCase
         (new QueryPlanner($this->adapter, 'SELECT'))->getPlan();
     }
 
-    public function testGetNumberOfRowsInspected()
+    public function testGetNumberOfRowsInspected(): void
     {
         $row1 = 1;
         $row2 = 2;
@@ -71,7 +78,7 @@ class QueryPlannerTest extends TestCase
         static::assertEquals($expected, $planner->getNumberOfRowsInspected());
     }
 
-    public function testGetNumberOfRowsInspectedDoesNotExceedMaxInt()
+    public function testGetNumberOfRowsInspectedDoesNotExceedMaxInt(): void
     {
         $row1 = PHP_INT_MAX;
         $row2 = 2;