+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- Type declarations have been added to all method parameters and return types
+  where possible.
 ### Changed
 - *BulkInsert*: **BC Break**: Private visibility for `fetchSql()` method. This
   shouldn't be used in regular implementations, as the SQL is used directly by